LogisticRegression worked
