+++ java/src/main/scala/com/nvidia/rapids/ml/RapidsLogisticRegression.scala
@@ -4,7 +4,7 @@ import org.apache.commons.logging.LogFactory
 import org.apache.spark.ml.util.Identifiable
 import org.apache.spark.sql.Dataset
 import org.apache.spark.ml.classification.{LogisticRegression, LogisticRegressionModel}
-import org.apache.spark.ml.rapids.{PythonRunner, RapidsMLFunction, RapidsUtils}
+import org.apache.spark.ml.rapids.{Fit, PythonRunner, RapidsMLFunction, RapidsUtils}
 
 
 class RapidsLogisticRegression(override val uid: String) extends LogisticRegression with RapidsEstimator {
@@ -18,22 +18,23 @@ class RapidsLogisticRegression(override val uid: String) extends LogisticRegress
 
     // Get the user-defined parameters and pass them to python process as a dictionary
     //    val params = this.paramMap
-
     val params = RapidsUtils.getUserDefinedParams(this)
     println(s"--------------------parameters of lr -------- ${params}")
 
-
     // TODO get the parameters (coefficients and intercepts) and construct the LogisticRegressionModel
-    val modelParameters = withResource(
-      new PythonRunner(estimatorName,
-        params,
+    withResource(
+      new PythonRunner(
+        Fit(estimatorName, uid, params),
         dataset.toDF,
+        RapidsUtils.buildLogisticRegressionModel,
         new RapidsMLFunction())) { runner =>
       runner.runInPython(useDaemon = false)
-    }
-    RapidsUtils.dummyLogisticRegressionModel
+    }.asInstanceOf[LogisticRegressionModel]
   }
 
-  override def estimatorName: String = "LogisticRegression"
 
+  /**
+   * The estimator name
+   */
+  override def estimatorName: String = "LogisticRegression"
 }
